move deserialize_move_struct to object.rs and flesh out display field
